Split flutter_service_worker_js to its own file.
Fixes https://github.com/flutter/flutter/issues/101828
